+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/nodes/Node.java
@@ -714,8 +714,8 @@ public abstract class Node implements NodeInterface, Cloneable {
             }
 
             @Override
-            public LanguageInfo createLanguage(Object vmObject, String id, String name, String version, Set<String> mimeTypes, boolean internal) {
-                return new LanguageInfo(vmObject, id, name, version, mimeTypes, internal);
+            public LanguageInfo createLanguage(Object vmObject, String id, String name, String version, Set<String> mimeTypes, boolean internal, boolean interactive) {
+                return new LanguageInfo(vmObject, id, name, version, mimeTypes, internal, interactive);
             }
 
             @Override